requirements-api.txt
# Core API dependencies
fastapi==0.117.1
uvicorn==0.37.0
pydantic==2.11.9

# Google Cloud dependencies
google-cloud-storage==2.18.0
google-cloud-logging==3.11.2

# Essential ML dependencies
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.2

# Configuration management
hydra-core==1.3.2
omegaconf==2.3.0

# Additional deployment dependencies
requests==2.31.0